#1167 Use retrofit also for capabilities endpoint
Fix API v0.2 compatibility
